requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dataclasses==0.6
future==0.18.2
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.2.0.34
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
urllib3==1.26.4
